bpf: dynamic allocation for bpf_verifier_env->subprog_info
At the moment only one hidden subprogram is allowed,
as bpf_verifier_env->subprog_info is allocated in advance as array of
fixed size. This patch removes the limitation by using dynamic memory
allocation for this array.

Spinning this out as a standalone patch from series [1],
to simplify the main series review in the future.

[1] https://lore.kernel.org/bpf/20241107175040.1659341-1-eddyz87@gmail.com/

Signed-off-by: Eduard Zingerman <eddyz87@gmail.com>