bpf: allow pointers for ARG_ANYTHING parameters of global functions
The goal of the commit is to simplify writing global functions that
convert their operands to read-only probed memory using
bpf_rdonly_cast(). Depending on the context such parameters might be
either pointers or scalars. E.g. here is how scalar parameters might
occur:

	...
	// `task` is `PTR_TO_BTF_ID`
	task = bpf_get_current_task_btf();
	// .name is defined as `char *`
	name = task->mm->exe_file->f_path.dentry->d_name.name;
	global_fn(name);

Verifier considers `name` to be a scalar because of how
`btf_struct_access` and `btf_struct_walk` process paths that end up
with pointers to non-struct types: such pointers are considered to be
scalar values.

This commit allows declaring global_fn as `global_fn(u64)` and pass
both pointers and scalars as in the example above as parameters of the
`global_fn`.

Suggested-by: Andrii Nakryiko <andrii@kernel.org>
Signed-off-by: Eduard Zingerman <eddyz87@gmail.com>