bpf: add bpf_features enum
This commit adds a kernel side enum for use in conjucntion with BTF
CO-RE bpf_core_enum_value_exists. The goal of the enum is to assist
with available BPF features detection.

Support for bpf_rdonly_cast to void* is the first feature listed in
the enum. Here is an example usage:

  if (bpf_core_enum_value_exists(enum bpf_features,
                                 BPF_FEAT_RDONLY_CAST_TO_VOID))
     ... bpf_rdonly_cast(..., 0) ...

Signed-off-by: Eduard Zingerman <eddyz87@gmail.com>